# pulsar-rest/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Browser4 WebDriver-Compatible API
  description: |
    A WebDriver-compatible HTTP API with selector-first extensions for browser automation.
    This API provides session management, navigation, element interaction, and event handling
    capabilities.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8182
    description: Local development server
...
paths:
...
  /session/{sessionId}:
    get:
      tags:
        - session
      summary: Get session details
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetails'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - session
      summary: Delete a session
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/element:
    post:
      tags:
        - element
      summary: Find element using strategy
      operationId: findElement
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindElementRequest'
      responses:
        '200':
          description: Element found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/elements:
    post:
      tags:
        - element
      summary: Find elements using strategy
      operationId: findElements
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindElementRequest'
      responses:
        '200':
          description: Elements found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/element/{elementId}/click:
    post:
      tags:
        - element
      summary: Click an element
      operationId: clickElement
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - $ref: '#/components/parameters/elementId'
      responses:
        '200':
          description: Element clicked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/element/{elementId}/value:
    post:
      tags:
        - element
      summary: Send keys to element
      operationId: sendKeysToElement
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - $ref: '#/components/parameters/elementId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendKeysRequest'
      responses:
        '200':
          description: Keys sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/element/{elementId}/attribute/{name}:
    get:
      tags:
        - element
      summary: Get element attribute
      operationId: getElementAttribute
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - $ref: '#/components/parameters/elementId'
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Attribute name
      responses:
        '200':
          description: Attribute value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/element/{elementId}/text:
    get:
      tags:
        - element
      summary: Get element text
      operationId: getElementText
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - $ref: '#/components/parameters/elementId'
      responses:
        '200':
          description: Element text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/event-configs:
    post:
      tags:
        - events
      summary: Create event configuration
      operationId: createEventConfig
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventConfig'
      responses:
        '200':
          description: Event config created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventConfigResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - events
      summary: Get event configurations
      operationId: getEventConfigs
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Event configurations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventConfigsResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/events:
    get:
      tags:
        - events
      summary: Get events
      operationId: getEvents
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Events list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/events/subscribe:
    post:
      tags:
        - events
      summary: Subscribe to events
      operationId: subscribeToEvents
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscribeRequest'
      responses:
        '200':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== PerceptiveAgent Endpoints ==========
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session identifier
    elementId:
      name: elementId
      in: path
      required: true
      schema:
        type: string
      description: Element identifier
  schemas:
...
    SessionDetails:
      type: object
      properties:
        value:
          type: object
          properties:
            sessionId:
              type: string
            url:
              type: string
            status:
              type: string
              enum: [active, paused, stopped]
            capabilities:
              type: object
              additionalProperties: true
    WebDriverResponse:
      type: object
      properties:
        value:
          description: Response value (can be null for success)
    ErrorResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            error:
              type: string
            message:
              type: string
            stacktrace:
              type: string
...
    FindElementRequest:
      type: object
      required:
        - using
        - value
      properties:
        using:
          type: string
          enum: [css selector, xpath, id, name, class name, tag name]
          description: Locator strategy
        value:
          type: string
          description: Locator value
    ElementRef:
      type: object
      properties:
        element-6066-11e4-a52e-4f735466cecf:
          type: string
          description: WebDriver element identifier
    ElementResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/ElementRef'
    ElementsResponse:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/ElementRef'
...
    SendKeysRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Text to send
    AttributeResponse:
      type: object
      properties:
        value:
          type: string
          description: Attribute value (can be null)
    TextResponse:
      type: object
      properties:
        value:
          type: string
          description: Element text content
...
    EventConfig:
      type: object
      required:
        - eventType
      properties:
        eventType:
          type: string
          description: Event type to listen for
        selector:
          type: string
          description: Optional CSS selector to scope events
        enabled:
          type: boolean
          default: true
    EventConfigResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            configId:
              type: string
            eventType:
              type: string
            enabled:
              type: boolean
    EventConfigsResponse:
      type: object
      properties:
        value:
          type: array
          items:
            type: object
            properties:
              configId:
                type: string
              eventType:
                type: string
              enabled:
                type: boolean
    Event:
      type: object
      properties:
        eventId:
          type: string
        eventType:
          type: string
        timestamp:
          type: integer
          format: int64
        data:
          type: object
          additionalProperties: true
    EventsResponse:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Event'
    SubscribeRequest:
      type: object
      required:
        - eventTypes
      properties:
        eventTypes:
          type: array
          items:
            type: string
          description: Event types to subscribe to
    SubscriptionResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            subscriptionId:
              type: string
            eventTypes:
              type: array
              items:
                type: string

    # ========== PerceptiveAgent Schemas ==========
...